Upload of current state of files
I've tried to universalise the main files so that they're not
dependant on other libraries I'm working on or anything site
specific. Currently there is some necessary server config required
to accomodate the style of ajax requests in the javascript.

Other than that it's a working prototype for standards compliant
browsers. It saves a single example page to a file as a sort of
database substitution.